+++ georocket-server/src/main/java/io/georocket/constants/ConfigConstants.java
@@ -52,6 +52,8 @@ public final class ConfigConstants {
   public static final String INDEX_MAX_BULK_SIZE = "georocket.index.maxBulkSize";
   public static final String INDEX_MAX_PARALLEL_INSERTS = "georocket.index.maxParallelInserts";
   public static final String INDEX_MAX_QUEUED_CHUNKS = "georocket.index.maxQueuedChunks";
+  public static final String INDEX_INDEXABLE_CHUNK_CACHE_MAX_SIZE = "georocket.index.indexableChunkCache.maxSize";
+  public static final String INDEX_INDEXABLE_CHUNK_CACHE_MAX_TIME_SECONDS = "georocket.index.indexableChunkCache.maxTimeSeconds";
   public static final String INDEX_ELASTICSEARCH_EMBEDDED = "georocket.index.elasticsearch.embedded";
   public static final String INDEX_ELASTICSEARCH_HOST = "georocket.index.elasticsearch.host";
   public static final String INDEX_ELASTICSEARCH_PORT = "georocket.index.elasticsearch.port";
@@ -72,6 +74,8 @@ public final class ConfigConstants {
   public static final int DEFAULT_INDEX_MAX_BULK_SIZE = 200;
   public static final int DEFAULT_INDEX_MAX_PARALLEL_INSERTS = 5;
   public static final int DEFAULT_INDEX_MAX_QUEUED_CHUNKS = 10000;
+  public static final long DEFAULT_INDEX_INDEXABLE_CHUNK_CACHE_MAX_SIZE = 1024L * 1024 * 64; // 64 MB
+  public static final long DEFAULT_INDEX_INDEXABLE_CHUNK_CACHE_MAX_TIME_SECONDS = 60;
 
   private ConfigConstants() {
     // hidden constructor